Upgrade to nom v4
There were two changes to nom v4 that had a major impact on this crate.

First, nom v4 uses the standard Result type instead of a custom enum.
That means that checking for a successful parse uses the is_ok method
instead of the is_done method.

Second, nom v4 is more aggressive about returning incomplete results.
This caused many tests to fail because the parser required one more
character to determine whether a keyword was finished or not. Work
around the problem by using nom v4's CompleteByteSlice type, which
indicates that the parser is never used with streaming input, and
therefore the end of the input string can be treated as EOF. This has
the added benefit of making all of the complete! macros redundant.